Fix annotations for multi configuration performance runs
[reviewed by kyle]

Annotations have to be attached to a specific series. Previously we just found
the first series listed and attached to that. However, with multiple
configurations the names of the series get changed, but the series that the
annotation got attached to did not. This moves where the annotations are
created to after where the configuration names are added and looks for a series
in the default configuration to attach to.

For us this means that if multiple configurations were run (say fifo and
qthreads) the annotations will get attached to a fifo series.

Eventually it would be nice to have a dummy line that the annotations get
attached to so that the annotations are displayed regardless of the what
configurations are hidden or not, but this gets us annotations in the
meantime.

This also turns off annotation generation if the numericX option is thrown as
our annotations are date based, and break if something other than dates are
used for the x-axis.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23659 3a8e244f-b0f2-452b-bcba-4c88e055c3ca