Initial support for Intel Phi
[reviewed by Greg]

This commit adds initial support for Intel Phi.  The support is
limited to Cray XC systems using PrgEnv-intel.  It's actually a very
small change to set the default GASNet conduit to 'mpi' rather than
'aries'.

The commit also adds a new README for using Chapel on Intel Phi.  It
will be changing quickly, but I thought it would be useful to have it
out for others to play around with.  The README includes TODOs for the
1.10 release.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23669 3a8e244f-b0f2-452b-bcba-4c88e055c3ca