Fixing _isSync bug where a single was thought to be a sync as well [reviewed by vass]
Vass discovered that _isSync was printing out true for singles as well as syncs.
This was because the FLAG_SYNC was attached to both and my compiler primitive
was looking for the presence of the flag rather than if it had that flag and was
not a single.  I fixed this by removing FLAG_SYNC from singles and replacing the
one instance where only FLAG_SYNC was used to now check for either FLAG_SYNC or
FLAG_SINGLE.  This resolved that portion of the bug.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23671 3a8e244f-b0f2-452b-bcba-4c88e055c3ca