Change how we check if dygraphs are ready to be synced along the x-axis
[trivial, not reviewed]

We used to have problems with syncing graphs along the x-axis before all the
graphs were fully rendered. The problem  was that if a graph we created but not
fully rendered you might be trying to set a property that wasn't actually
initialized yet. To get around that we started tracking how many graphs were
fully rendered and checked that number against the total number of graphs
created before syncing along the x-axis. However, for some reason a graph will
occasionally get created but it never renders itself (a problem to figure out
for another time) and because of that we would have x graphs created but only
x-1 graphs rendered no matter how long you waited so you could never sync all
your graphs.

This just makes it so that each graph has its own isReady property so that even
if not all graphs have actually rendered themselves you can still sync the
ones that have.

This problem usually manifests itself by preventing syncing of graphs along the
x-axis and if you scroll down far enough you see a white spot where a graph
should be. This will not eliminate that white spot (that is an un-rendered
graph) but will allow you to sync the other graphs that have rendered along the
x-axis


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23677 3a8e244f-b0f2-452b-bcba-4c88e055c3ca