Fix for new cygwin 'uname' behavior
In a recent cygwin version, it seems that uname has started returning strings
like 'CYGWIN_NT-6.1-WOW64' rather than simply 'CYGWIN'.  This commit changes
the 'chpl_platform.py' script to change any uname[0] values that start with
'cygwin' to simply 'cygwin'.  We don't currently have any need (that we're
aware of) to distinguish between different windows versions, so this seems
reasonable.

If we do need to do so in the future, we could preserve the uname string
(which was the old behavior) and then change checks against 'cygwin' to check
for 'cygwin*' in other places (like chpl_make.py).






git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23678 3a8e244f-b0f2-452b-bcba-4c88e055c3ca