Remove setting QT_GUARD_PAGES and QT_WORKER_UNIT from common-qthreads
[ok'd by locale model team]

common-qthreads is sourced for nightly qthreads testing. It should be testing
the "default" configurations but we currently set QT_GUARD_PAGES=true and
QT_WORKER_UNIT=core. It happens that guard_pages are on by default and the
default worker unit is core so these have no real affect as we are just
re-setting part of the default ocnfiguration, but it makes sense to remove them
anyways. They were originally set because the default worker unit wasn't core,
and we wanted to be explicit about setting guard pages, but they are not
needed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23679 3a8e244f-b0f2-452b-bcba-4c88e055c3ca