Switch the hwloc default version from 1.7.2 to 1.9.
[Reviewed by Sung and Brad.]

The actual version number change is the one line in Makefile.include.
The other changes, in Makefile, are:
- Remove the code that unpacks the tarfile.  We don't need this for
  hwloc 1.9, because as of that release our repository already contains
  the unpacked tar file contents.
- Add some touches so that the autoconf and automake steps don't run.
  These are just like the ones in our third-party/qthreads/Makefile.

In the original version of this patch I'd left out the sleep commands
from between the touch commands in that second part.  (I copied that
touch logic from third-party/qthreads/Makefile, but it originated in
GASNet's Bootstrap file.)  It seemed to me that they shouldn't be
required.  But after a conversation about that with the reviewers I
decided I might be suffering from a lack of imagination in that regard,
so I put them back in.  Seeing if they can be removed is unrelated to
this default version change anyway, and will make a good rainy-day task.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23682 3a8e244f-b0f2-452b-bcba-4c88e055c3ca