Enable fifo/qthreads graph generation for chap03
[trivial, not reviewed]

Qthreads graphs were only getting generated for the qthreads run. This was so
that the first night of testing wouldn't have a bunch of errors reported. This
resulted in just fifo results being displayed until the qthreads testing was
done. Now that we have qthreads data, the fifo run won't complain about trying
to create the combined graphs for fifo/qthreads.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23692 3a8e244f-b0f2-452b-bcba-4c88e055c3ca