First cut at a filerator (iterator that yields file[name]s)
I've been splashing around today with writing a (serial) iterator that
yields files.  The current interface takes a starting directory, a
boolean saying whether or not to traverse the hierarchy recursively,
and a boolean saying whether or not to consider dotfiles.  So far, it
seems to be working nicely for me.  This commit adds .execopts to try
the 2 x 2 set of boolean values and a little directory structure with
dotfiles and non- to test behavior.

I'm checking this in as a NOTEST for now because the order of the
files being yielded is nondeterministic, based on the internals of
one's filesystem (I think).  You'd need to sort to avoid this, which
would require manifesting the filenames in memory (for better or
worse).  The technique I'm relying on to determine if a filename is a
directory in the recursive case is also non-portable.

Arguably, I also shouldn't be yielding the directory names... that
only occurred to me just now.

Now that I'm comfortable with the C interface, the next steps are to
study the Python interface for reference, propose something that seems
like a good fit for Chapel, and circulate that proposal for comment.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23699 3a8e244f-b0f2-452b-bcba-4c88e055c3ca