Fix a long standing issue with whitebox testing non-PrgEnv gnu compiler.
Previously, nightly.chpbld01 was loading a module name "gnu", which is not the
compiler module associated PrgEnv-gnu. We really want to load gcc. gnu is some
legacy module, which does not do anything on Linux -- it appears to provide
some GNU tools on Sun systems.

This does not change nightly.chpbld01, just the newly added
util/cron/test-white-examples.bash.

git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23703 3a8e244f-b0f2-452b-bcba-4c88e055c3ca