Removing ptrans and stream perfnumtrials
[ok'd by lydia]

Removing these two perfnumtrials to get multiple trials on stream and ptrans.
Ptrans was originally set to one because it took ~30 minutes on chap04, but
that was fixed with r23570.

The stream one is being removed for similar reasons as stream-ep's from r23206.

If it matters: these were the only two perfnumtrials in the repo so the only
tests that don't have multiple trials are the ones with timeouts (which is
currently just threadring I think.) 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23708 3a8e244f-b0f2-452b-bcba-4c88e055c3ca