Add support for logical negation of integral types
Prior to this patch, the ! operator was only supported for bool types. This
patch extends that support to also support integral types with the obvious
interpretation (zero => true, non-zero => false).  This implementation adds
that support in both param and non-param versions. It also adds a few tests
to lock in the behavior and documentation to the spec for the non-param
versions (the existing param version did not seem to be documented in the
spec, so I followed suit here for now, and added a TODO to the spec TODO to
figure out how we want to document these in the future).

I believe that the rationale for the past behavior has probably been "it's
what C# does."  I believe C#'s rationale is to avoid making typos like the
following legal programs:

        if (!(x=5)) ...

where the user probably meant to type:

        if (!(x==5)) ...

Chapel prevents these typos another way -- by only permitting assignments at
the statement level, so that may be a reason to feel OK diverging from C#'s
practice.


Note that Chapel does permit integer expressions in conditionals, e.g.:

        var x: int;
        ...

        if (x) ...

which is why it seemed suprising to me to find that we do not support:

        if (!x)

Proposing this on chapel-developers yesterday, there were no objections, and
more positive response than most requests for comment generate.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23711 3a8e244f-b0f2-452b-bcba-4c88e055c3ca