Change the return type of is_c_nil from c_int to bool
[reviewed by lydia]

This commit changes the return type of is_c_nil from c_int to bool. I
reviewed the concept with Michael (who wrote is_c_nil() originally)
before implementing.  I believe that is_c_nil returns a c_int simply
because it was wrapping a C comparison operator that returned an int;
making it a bool seems more natural for Chapel's sake and removes a
case where we rely on C types in something that an end-user or library
writer might want to use.

I implemented this by casting the result to chpl_bool in the C routine
that implements c_is_bool().  To do this, I had to move the definition
of chpl_bool up in the file.

I added a test to lock in this capability.

The reason I ran into this is that I was writing code of the form:

        if (!is_c_nil(x))

only to find that it wouldn't work due to our lack of support for ! on
int types, which led me to be surprised that is_c_nil returned an int.
In another patch, I've extended support of ! to include integers; this
commit solves the same issue in a different way and is not strictly
necessary with the previous, but feels like the cleaner/simpler
interface to support.

I also simplified existing uses of is_c_nil() to avoid the need to
compare against integers, based on this change.


Note for discussion: I think I would've called Chapel's notion of C's
'NULL' 'c_NULL' rather than 'c_nil', much like we call C's 'double'
'c_double' rather than 'c_real'.  Put another way, I worry that a C
programmer using Chapel would not find/guess 'c_nil' as quickly as
they would 'c_NULL' (I can never find it, anyway).  Thoughts?




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23713 3a8e244f-b0f2-452b-bcba-4c88e055c3ca