Minor style cleanup to my previous !-for-ints commit
[proposed by hilde, who reviewed the previous patch]

This commit changes the implementation of the param version of !() on
integers to use the same implementation as the non-param versions
(comparing to zero, rather than using a primitive), for the purpose of
making the two versions more similar and reducing the reliance on
primitives.

This re-raises a topic that was discussed on the Cray chatroom this
week about what it would take to write a function that could query the
param-ness of its arguments (was polymorphic over param-ness, say) and
optionally assert its own param-ness based on that information.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23714 3a8e244f-b0f2-452b-bcba-4c88e055c3ca