Improvements to filerator to make it testable
This commit changes the filerator to give it an option to skip over .svn
directories (for the purposes of exploration and testing; not suggesting
that this should necessarily be part of the final interface), and adds a
sorterator -- an iterator that takes another iterator, evaluates it, sorts
the results, and yields those -- to make the output deterministic.  This
should be sufficient to make this test portable and useful, so I'm removing
the NOTEST for now and will test in a few places to be sure.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23715 3a8e244f-b0f2-452b-bcba-4c88e055c3ca