Add macroeconomics benchmark to test suite.
A recently published paper compared the performance of several programming
languages against a benchmark that is relevant to macroeconomics. The benchmark
implements, "the basic stochastic neoclassical growth model, which is the
foundation of much work in modern macroeconomics." Languages included: C++,
Fortran, Julia, Python (with Numpy and Numba), Java, Matlab, R, and
Mathematica.

The article that led me to this:

  http://marginalrevolution.com/marginalrevolution/2014/07/a-comparison-of-programming-languages-in-economics.html

The actual paper:

  http://economics.sas.upenn.edu/~jesusfv/comparison_languages.pdf

Source code for implementations in C++, Python, Julia, etc:

  https://github.com/jesusfv/Comparison-Programming-Languages-Economics

The good news is that this Chapel version is ~2.2x slower than the C++ version
(the top performer in the paper) on my machine. I suspect this will vary quite
a bit depending on the the natural degree of parallelism on the machine running
the benchmark. For reference, the paper says that Java and Julia were ~2-3x
slower than the C++ version, so this version is competitive.

The Chapel version here is based off the Java version. I made it Chapel-like
where it was convenient and left some comments where things could be
improved. It would be interesting to port the Julia or Python versions, as
well.

One issue that I noticed is the precision of the Chapel result is off
slightly. I couldn't tell if this was due to an algorithmic difference or if
our reals were "less precise" somehow than Java doubles, as an example. Seems
more likely that there is an algorithmic error. In any case, I marked it as
NOTEST to avoid issues since I have not tested it across platforms.

I added a perfkeys file to extract the elapsed time that was used in the study
and verify the "My check" result.

git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23717 3a8e244f-b0f2-452b-bcba-4c88e055c3ca