Don't try to rebuild the hwloc doxygen docs.
[Reviewed conversationally with Elliot.]

Because we've stored the contents of the hwloc tarball into our
repository and we build with files pulled back out of the repo, we've
lost the file timestamps that are in the tarball.  As a result, make(1)
sometimes thinks it needs to rebuild things when in fact it doesn't.
Among these things are the doxygen docs; lately some of us have seen
attempts to rebuild those.  Unfortunately something about our use of a
separate build directory prevents the doxygen doc rebuild step from
working properly.  Combining the timestamp problem and the separate
build directory problem results in sporadic hwloc build failures.

This change touches $(HWLOC_SUBDIR)/README so that it is as new as its
prerequisite $(HWLOC_SUBDIR)/doc/doxygen-doc/hwloc.tag, preventing
make from trying to do the doxygen doc rebuild.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23720 3a8e244f-b0f2-452b-bcba-4c88e055c3ca