Update util/cron/ script to use current working directory as CHPL_HOME.
Previously these scripts relied on ~/chapel as the main repo and would default
to using it as CHPL_HOME. The new testing system guarantees that the current
working directory is a clean, up-to-date working copy so we can use it.

Also, a couple of scripts are added to the perl compilation check in smokeTest
and some additional files are now ignored in .gitignore.

git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23721 3a8e244f-b0f2-452b-bcba-4c88e055c3ca