Update whitebox testing scripts to use new testing system.
Removes the old chpbld01-02 (aka whitebox) testing scripts, which used for
loops to test a bunch of different host and target compilers with and without a
programming environment.

The new testing system supports configuration matrices, which is how the
various compilers and programming environment are now expressed. The scripts in
this repo know how to setup a single configuration based on some environment
variables and then run nightly.

git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23724 3a8e244f-b0f2-452b-bcba-4c88e055c3ca