To for_formals_actuals, add a test that a PRIM_VIRTUAL_METHOD_CALL gets the right number of arguments.
Added a clause to the for_formals_actuals macro, to also test that the argument list is
the right length when a virtual function (as opposed to a statically-bound function) is
being called.

This caused two test cases to fail:
  classes/deitz/dispatch/test_generic_where.chpl
  distributions/diten/Dimensional.chpl
because param and type arguments were being removed from calls only if the function called
was statically-bound.  I added code to removeRandomCalls(), to remove these arguments from
calls to virtual (dynamically-bound) functions as well.  After that, the tests passed.

[tested on / (std]
[reviewed by noakes]




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23725 3a8e244f-b0f2-452b-bcba-4c88e055c3ca