Remove "string_normalize" primitive
[reviewed by David]

This patch removes the "string_normalize" primitive added in r21443.
It is no longer needed, as Chapel strings are no longer passed to or
returned from extern functions.  The definition and codegen for the
primitive is removed from the compiler, and all uses are removed from
the modules and 3 test programs.

This patch also changes the remaining extern calls using Chapel
strings to use c_string instead (these were missed in r23558), minus
the one in HDFS (Tim already has a patch to fix that).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23730 3a8e244f-b0f2-452b-bcba-4c88e055c3ca