Put compile-time configuration variables into their own generated file.
[Reviewed by Sung.]

This puts the compile-time configuration info into a separate file,
chpl_compilation_config.h.  This is #include'd into chpl__header.h for
the user program, so in that regard there is no change.  But it can also
be #include'd into a launcher, and then launchers could also know the
compile-time configuration.

This also enables doing something like an "--about" execution-time
option, which would describe the compile-time configuration for a user
program.  We could do that before, but if there was a launcher we would
have had to implement "--about" in the *_real user executable.  Now we
can do it in the launcher, which would be more elegant.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23738 3a8e244f-b0f2-452b-bcba-4c88e055c3ca