Pull correct value for final check in RBC benchmark.
Previously this was using `mPolicyFunction[999, 2]`, which is correct if the
domain has zero based indexes. Chapel, however, uses one based indexes, so it
is now correctly `mPolicyFunction[1000, 3]` and the value matches the expected!

The .perfkeys and .good files are also updated to check on the correct value. I
removed the NOTEST file so it will now be included in testing.

git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23740 3a8e244f-b0f2-452b-bcba-4c88e055c3ca