Add RBCElegant, which uses dotProduct and transpose function.
This makes the top of the while loop look a lot nicer, but as expected it
causes a significant performance decrease. This is due to allocating a new
domain and arrary each time those functions are called.

RBCElegant has comparable performance to the Python with Numpy version on my
mac.