Set CHPL_HOST_COMPILER=gnu when testing gnu.
I accidentally switched this to gcc when I fixed which "gnu" module we were
loading in r23724.

This also fixes the issues qio issues we saw recently when building the
compiler. Since the compiler did not recognize "gcc" as a valid compiler,
several key flags were left off when compiling the sources.

git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23743 3a8e244f-b0f2-452b-bcba-4c88e055c3ca