Revert r23702: Fix lldb support
[trivial, not reviewed]

I originally made this change to make --lldb work correctly after some
changes to how argument parsing worked. As it turns out I was really just
putting in a workaround for a bug that has since been fixed. After reverting
this change, arguments with a double dash should be properly passed on to
chpl when using --lldb.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23763 3a8e244f-b0f2-452b-bcba-4c88e055c3ca