Don't override the user's QTHREAD_WORKER_UNIT with our QT_WORKER_UNIT.
[Reviewed by Elliot.]

Qthreads supports both QT_* and QTHREAD_* forms of its environment
variables.  We want to run with the Qthreads worker unit value set to
"core".  In the setenv(3) that sets that our final argument is 0, so we
won't override a user's pre-existing setting for QT_WORKER_UNIT.  But if
they set QTHREAD_WORKER_UNIT instead, then we will override it, because
the QT_* env vars override the QTHREAD_* ones within Qthreads (though
this is not documented and one shouldn't depend on it).  The fix is to
not set QT_WORKER_UNIT if QTHREAD_WORKER_UNIT is already set.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23769 3a8e244f-b0f2-452b-bcba-4c88e055c3ca