Fix detection of missing return types to avoid a false positive case [reviewed by hilde]
It was discovered that the presence of a variable of extern type in a function
would remove the flag FLAG_SPECIFIED_RETURN_TYPE from the function it lived in,
even if this variable was not actually the variable returned by that function.
To fix this, I removed the code that removed the flag and added a test to
checkReturnPaths that would handle the extern case we were originally trying to
avoid.  I also added two tests which had failed under the previous state but now
pass.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23771 3a8e244f-b0f2-452b-bcba-4c88e055c3ca