Include string leaks in leaking programs reported.
This patch adds a new option "-show_strings" to the analyzeMemLeaksLog script, which
includes string-related memory descriptors when tallying which programs leak.  
It does not affect the number of bytes reported as leaked.  The difference will be that it
will now list those programs that leak only string-related data, and include them in the
total number of programs that leak.

The nightly cron script was modified to run analyzeMemLeaksLog with the new -show_strings
option..

[tested locally]
[not reviewed]



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23785 3a8e244f-b0f2-452b-bcba-4c88e055c3ca