Fix externProcInClassShouldBeError.chpl and change its future messages
This test should have been changed with the proc = signature changes
(no .bad file made it pretty much impossible to spot).

After updating, the test no longer fails in the way it originally did
(internal error).  Not sure what fixed the problem along the way, but
it now fails in a somewhat less confusing way (during C compile), so
I've turned this into an error message future (or maybe it should be a
"warning message" future).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23789 3a8e244f-b0f2-452b-bcba-4c88e055c3ca