Refactor insertAutoDestroyCalls() in callDestructors.cpp
Partitioned the business logic in insertAutoDestroyCalls() into a few
smaller functions

   a) to make it a little less opaque

   b) to support pending mods for the current Interpreter task


This submit does not contain any changes to the business logic.



Rationale (Short Version)
=========================

An effort to make "a small change" to the parser's AST for most
procedures caused just over 100 paratest failures.  I studied one
failure in some detail and determined that it could be fixed by an
"incomplete change" to insertAutoDestroyCalls().

To my surprise, and perhaps even partial dismay, my poorly conceived
change eliminated all of the failures.

A little more study using the memleaks option to start_test and
paratest indicated that my change appeared to cause a modest bump, 
about 10% or so, in memory leakage.

I'd like to spend a little more time on this issue to determine if
I can develop a more complete fix that enables my desired trivial
transformation without introducing an additional memory leak.

I believed this would be easier if the code were a little more
transparent.




insertAutoDestoryCalls() walks the statements of every BlockStmt
and performs 3 important tests on each stmt. The first test determines
whether the statement defines a variable that needs to be auto-destroyed
and the other two tests determine whether the statement includes control
flow that leaves the BlockStmt and requires the insertion of autoDestroy
code.


I have partitioned the logic in to several functions to better match
the apparent intent.  This is in preparation for some minor improvements
to the code and then a second attempt to implement a correct fix for the
issue that drove me in to this function.





Compiled:   STD on clang.  STD/LLVM on gcc

Tested:     paratest
            start_test with -memleaks on release/examples
            paratest   with -memleaks

Review:     Tom






git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23801 3a8e244f-b0f2-452b-bcba-4c88e055c3ca